+++ openfisca_survey_manager/scenarios/reform_scenario.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 
 
-from openfisca_survey_manager.scenarios.abstract_scenario import AbstractSurveyScenario, diagnose_variable_mismatch, init_variable_in_entity
+from openfisca_survey_manager.scenarios.abstract_scenario import AbstractSurveyScenario
 # from openfisca_core import periods
 # from openfisca_survey_manager.simulations import Simulation  # noqa analysis:ignore
 # from openfisca_core.simulation_builder import SimulationBuilder
@@ -68,45 +68,6 @@ class ReformScenario(AbstractSurveyScenario):
             simulation = self.simulation
 
         return simulation.adaptative_calculate_variable(variable, period = period)
-
-    # def calibrate(self, period: int = None, target_margins_by_variable: dict = None, parameters: dict = None, target_entity_count: float = None):
-    #     """Calibrate the scenario data.
-
-    #     Args:
-    #         period (int, optionnal): Period of calibration. Defaults to scenario.year
-    #         target_margins_by_variable (dict, optional): Variable targets margins. Defaults to None.
-    #         parameters (dict, optional): Calibration parameters. Defaults to None.
-    #         total_population (float, optional): Total population target. Defaults to None.
-    #     """
-    #     survey_scenario = self
-
-    #     if period is None:
-    #         assert survey_scenario.period is not None
-    #         period = survey_scenario.period
-
-    #     if parameters is not None:
-    #         assert parameters['method'] in ['linear', 'raking ratio', 'logit'], \
-    #             "Incorect parameter value: method should be 'linear', 'raking ratio' or 'logit'"
-    #         if parameters['method'] == 'logit':
-    #             assert parameters['invlo'] is not None
-    #             assert parameters['up'] is not None
-    #     else:
-    #         parameters = dict(method = 'logit', up = 3, invlo = 3)
-
-    #     # TODO: filtering using filtering_variable_by_entity
-    #     for simulation in [survey_scenario.simulation, survey_scenario.baseline_simulation]:
-    #         if simulation is None:
-    #             continue
-    #         calibration = Calibration(
-    #             simulation,
-    #             target_margins_by_variable,
-    #             period,
-    #             target_entity_count = target_entity_count,
-    #             parameters = parameters,
-    #             # filter_by = self.filter_by,
-    #             )
-    #         calibration.calibrate(inplace = True)
-    #         simulation.calibration = calibration
 
     def compute_aggregate(self, variable = None, aggfunc = 'sum', filter_by = None, period = None, use_baseline = False,
             difference = False, missing_variable_default_value = np.nan, weighted = True, alternative_weights = None):
@@ -376,276 +337,3 @@ class ReformScenario(AbstractSurveyScenario):
             merge = merge,
             id_variable_by_entity_key = id_variable_by_entity_key,
             )
-
-    def custom_input_data_frame(self, input_data_frame, **kwargs):
-        """Customize input data frame.
-
-        Args:
-          input_data_frame: original input data frame.
-          kwargs: keyword arguments.
-        """
-        pass
-
-    def inflate(self, inflator_by_variable = None, period = None, target_by_variable = None):
-        assert inflator_by_variable or target_by_variable
-        assert period is not None
-        inflator_by_variable = dict() if inflator_by_variable is None else inflator_by_variable
-        target_by_variable = dict() if target_by_variable is None else target_by_variable
-        self.inflator_by_variable = inflator_by_variable
-        self.target_by_variable = target_by_variable
-
-        assert self.simulation is not None
-        for use_baseline in [False, True]:
-            if use_baseline:
-                simulation = self.baseline_simulation
-            else:
-                assert self.simulation is not None
-                simulation = self.simulation
-                if (self.simulation == self.baseline_simulation):  # Avoid inflating two times
-                    continue
-
-            if simulation is None:
-                continue
-
-            tax_benefit_system = self.tax_benefit_system
-            for variable_name in set(inflator_by_variable.keys()).union(set(target_by_variable.keys())):
-                assert variable_name in tax_benefit_system.variables, \
-                    "Variable {} is not a valid variable of the tax-benefit system".format(variable_name)
-                if variable_name in target_by_variable:
-                    inflator = inflator_by_variable[variable_name] = \
-                        target_by_variable[variable_name] / self.compute_aggregate(
-                            variable = variable_name, use_baseline = use_baseline, period = period)
-                    log.info('Using {} as inflator for {} to reach the target {} '.format(
-                        inflator, variable_name, target_by_variable[variable_name]))
-                else:
-                    assert variable_name in inflator_by_variable, 'variable_name is not in inflator_by_variable'
-                    log.info('Using inflator {} for {}.  The target is thus {}'.format(
-                        inflator_by_variable[variable_name],
-                        variable_name, inflator_by_variable[variable_name] * self.compute_aggregate(
-                            variable = variable_name, use_baseline = use_baseline, period = period)
-                        ))
-                    inflator = inflator_by_variable[variable_name]
-
-                array = simulation.calculate_add(variable_name, period = period)
-                assert array is not None
-                simulation.delete_arrays(variable_name, period = period)  # delete existing arrays
-                simulation.set_input(variable_name, period, inflator * array)  # insert inflated array
-
-    def init_from_data(self, calibration_kwargs = None, inflation_kwargs = None,
-            rebuild_input_data = False, rebuild_kwargs = None, data = None, memory_config = None, use_marginal_tax_rate = False,
-            config_files_directory = default_config_files_directory):
-        """Initialise a survey scenario from data.
-
-        Args:
-          rebuild_input_data(bool):  Whether or not to clean, format and save data. Take a look at :func:`build_input_data`
-          data(dict): Contains the data, or metadata needed to know where to find it.
-          use_marginal_tax_rate(bool): True to go into marginal effective tax rate computation mode.
-          calibration_kwargs(dict):  Calibration options (Default value = None)
-          inflation_kwargs(dict): Inflations options (Default value = None)
-          rebuild_input_data(bool): Whether to rebuild the data (Default value = False)
-          rebuild_kwargs:  Rebuild options (Default value = None)
-          config_files_directory:  Directory where to find the configuration files (Default value = default_config_files_directory)
-        """
-        # When not ``None``, it'll try to get the data for *period*.
-        if data is not None:
-            data_year = data.get("data_year", self.period)
-
-        self._set_id_variable_by_entity_key()
-        self._set_role_variable_by_entity_key()
-        self._set_used_as_input_variables_by_entity()
-
-        # When ``True`` it'll assume it is raw data and do all that described supra.
-        # When ``False``, it'll assume data is ready for consumption.
-        if rebuild_input_data:
-            if rebuild_kwargs is not None:
-                self.build_input_data(year = data_year, **rebuild_kwargs)
-            else:
-                self.build_input_data(year = data_year)
-
-        debug = self.debug
-        trace = self.trace
-
-        if use_marginal_tax_rate:
-            assert self.varying_variable in self.tax_benefit_system.variables
-
-        # Inverting reform and baseline because we are more likely
-        # to use baseline input in reform than the other way around
-        if self.baseline_tax_benefit_system is not None:
-            self.new_simulation(debug = debug, data = data, trace = trace, memory_config = memory_config,
-                use_baseline = True)
-            if use_marginal_tax_rate:
-                self.new_simulation(debug = debug, data = data, trace = trace, memory_config = memory_config, use_baseline = True,
-                    marginal_tax_rate_only = True)
-
-        # Note that I can pass a :class:`pd.DataFrame` directly, if I don't want to rebuild the data.
-        self.new_simulation(debug = debug, data = data, trace = trace, memory_config = memory_config)
-
-        if use_marginal_tax_rate:
-            self.new_simulation(debug = debug, data = data, trace = trace, memory_config = memory_config, marginal_tax_rate_only = True)
-
-        self.set_weight_variable_by_entity()
-
-        if calibration_kwargs is not None:
-            assert set(calibration_kwargs.keys()).issubset(set(
-                ['target_margins_by_variable', 'parameters', 'total_population']))
-
-        if inflation_kwargs is not None:
-            assert set(inflation_kwargs.keys()).issubset(set(['inflator_by_variable', 'target_by_variable', 'period']))
-
-        if calibration_kwargs:
-            self.calibrate(**calibration_kwargs)
-
-        if inflation_kwargs:
-            self.inflate(**inflation_kwargs)
-
-    def init_entity_structure(self, tax_benefit_system, entity, input_data_frame, builder):
-        """Initialize sthe simulation with tax_benefit_system entities and input_data_frame.
-
-        Args:
-          tax_benefit_system(TaxBenfitSystem): The TaxBenefitSystem to get the structure from
-          entity(Entity): The entity to initialize structure
-          input_data_frame(pd.DataFrame): The input
-          builder(Builder): The builder
-
-        """
-        id_variables = [
-            self.id_variable_by_entity_key[_entity.key] for _entity in tax_benefit_system.group_entities]
-        role_variables = [
-            self.role_variable_by_entity_key[_entity.key] for _entity in tax_benefit_system.group_entities]
-
-        if entity.is_person:
-            for id_variable in id_variables + role_variables:
-                assert id_variable in input_data_frame.columns, \
-                    "Variable {} is not present in input dataframe".format(id_variable)
-
-        input_data_frame = self.filter_input_variables(input_data_frame = input_data_frame)
-
-        ids = range(len(input_data_frame))
-        if entity.is_person:
-            builder.declare_person_entity(entity.key, ids)
-            for group_entity in tax_benefit_system.group_entities:
-                _key = group_entity.key
-                _id_variable = self.id_variable_by_entity_key[_key]
-                _role_variable = self.role_variable_by_entity_key[_key]
-                group_population = builder.declare_entity(_key, input_data_frame[_id_variable].drop_duplicates().sort_values().values)
-                builder.join_with_persons(
-                    group_population,
-                    input_data_frame[_id_variable].astype('int').values,
-                    input_data_frame[_role_variable].astype('int').values,
-                    )
-
-    def init_entity_data(self, entity, input_data_frame, period, simulation):
-        used_as_input_variables = self.used_as_input_variables_by_entity[entity.key]
-        diagnose_variable_mismatch(used_as_input_variables, input_data_frame)
-        input_data_frame = self.filter_input_variables(input_data_frame = input_data_frame)
-
-        for column_name, column_serie in input_data_frame.items():
-            variable_instance = self.tax_benefit_system.variables.get(column_name)
-            if variable_instance is None:
-                continue
-
-            if variable_instance.entity.key != entity.key:
-                log.info("Ignoring variable {} which is not part of entity {} but {}".format(
-                    column_name, entity.key, variable_instance.entity.key))
-                continue
-            init_variable_in_entity(simulation, entity.key, column_name, column_serie, period)
-
-    def init_simulation_with_data_frame(self, tax_benefit_system, input_data_frame, period, builder):
-        """Initialize the simulation period with current input_data_frame for an entity if specified."""
-        used_as_input_variables = self.used_as_input_variables
-        id_variable_by_entity_key = self.id_variable_by_entity_key
-        role_variable_by_entity_key = self.role_variable_by_entity_key
-
-        diagnose_variable_mismatch(used_as_input_variables, input_data_frame)
-
-        id_variables = [
-            id_variable_by_entity_key[_entity.key] for _entity in tax_benefit_system.group_entities]
-        role_variables = [
-            role_variable_by_entity_key[_entity.key] for _entity in tax_benefit_system.group_entities]
-
-        for id_variable in id_variables + role_variables:
-            assert id_variable in input_data_frame.columns, \
-                "Variable {} is not present in input dataframe".format(id_variable)
-
-        input_data_frame = self.filter_input_variables(input_data_frame = input_data_frame)
-
-        index_by_entity_key = dict()
-
-        for entity in tax_benefit_system.entities:
-            self.init_entity_structure(tax_benefit_system, entity, input_data_frame, builder)
-
-            if entity.is_person:
-                continue
-
-            else:
-                index_by_entity_key[entity.key] = input_data_frame.loc[
-                    input_data_frame[role_variable_by_entity_key[entity.key]] == 0,
-                    id_variable_by_entity_key[entity.key]
-                    ].sort_values().index
-
-        for column_name, column_serie in input_data_frame.items():
-            if role_variable_by_entity_key is not None:
-                if column_name in role_variable_by_entity_key.values():
-                    continue
-
-            if id_variable_by_entity_key is not None:
-                if column_name in id_variable_by_entity_key.values():
-                    continue
-
-            simulation = builder.build(tax_benefit_system)
-            entity = tax_benefit_system.variables[column_name].entity
-            if entity.is_person:
-                init_variable_in_entity(simulation, entity.key, column_name, column_serie, period)
-            else:
-                init_variable_in_entity(simulation, entity.key, column_name, column_serie[index_by_entity_key[entity.key]], period)
-
-        return simulation
-
-    # def memory_usage(self, use_baseline = False):
-    #     if use_baseline:
-    #         simulation = self.baseline_simulation
-    #     else:
-    #         simulation = self.simulation
-
-    #     memory_usage_by_variable = simulation.get_memory_usage()['by_variable']
-    #     try:
-    #         usage_stats = simulation.tracer.usage_stats
-    #     except AttributeError:
-    #         log.warning("The simulation trace mode is not activated. You need to activate it to get stats about variable usage (hits).")
-    #         usage_stats = None
-    #     infos_lines = list()
-
-    #     for variable, infos in memory_usage_by_variable.items():
-    #         hits = usage_stats[variable]['nb_requests'] if usage_stats else None
-    #         infos_lines.append((
-    #             infos['total_nb_bytes'],
-    #             variable, "{}: {} periods * {} cells * item size {} ({}) = {} with {} hits".format(
-    #                 variable,
-    #                 infos['nb_arrays'],
-    #                 infos['nb_cells_by_array'],
-    #                 infos['cell_size'],
-    #                 infos['dtype'],
-    #                 humanize.naturalsize(infos['total_nb_bytes'], gnu = True),
-    #                 hits,
-    #                 )
-    #             ))
-    #     infos_lines.sort()
-    #     for _, _, line in infos_lines:
-    #         print(line.rjust(100))  # noqa analysis:ignore
-
-    def set_input_data_frame(self, input_data_frame):
-        """Set the input dataframe.
-
-        Args:
-          input_data_frame (pd.DataFrame): Input data frame
-
-        """
-        self.input_data_frame = input_data_frame
-
-    def set_weight_variable_by_entity(self, weight_variable_by_entity = None):
-        if weight_variable_by_entity is not None:
-            self.weight_variable_by_entity = weight_variable_by_entity
-
-        for simulation in [self.simulation, self.baseline_simulation]:
-            if simulation is not None:
-                simulation.weight_variable_by_entity = self.weight_variable_by_entity
